remade footer with focus on discord funnel, added invite button to top right nav, installed a few shadcn ui components
